fix: 继续优化项目，符合规范，补全注释

+++ src/main/java/game605/controller/ImgCommentController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.*;
  * @author Koyou
  * @version 1.0.0
  * @className CommentController
- * @description TODO
+ * @description 图片评论功能
  * @since 2024/7/11 16:24
  */
 @RestController
@@ -44,7 +44,7 @@ public class ImgCommentController {
      * 获取图片评论列表
      */
     @GetMapping("/getCommentList")
-    public ResponseResult getImgCommentList(@RequestBody Long imgId){
+    public ResponseResult getImgCommentList(@RequestBody Integer imgId){
         return ResponseResult.success(commentService.getImgCommentList(imgId));
     }
